refactor: Delete 할 Station이 없을 때 BadRequest(400)로 응답하도록 수정

+++ src/main/java/wooteco/subway/station/application/StationService.java
@@ -2,7 +2,9 @@ package wooteco.subway.station.application;
 
 import java.util.List;
 import java.util.stream.Collectors;
+import org.springframework.http.HttpStatus;
 import org.springframework.stereotype.Service;
+import wooteco.subway.exception.HttpException;
 import wooteco.subway.station.dao.StationDao;
 import wooteco.subway.station.domain.Station;
 import wooteco.subway.station.dto.StationRequest;
@@ -34,6 +36,10 @@ public class StationService {
     }
 
     public void deleteStationById(Long id) {
+        Station foundStation = stationDao.findById(id);
+        if (foundStation == null) {
+            throw new HttpException(HttpStatus.BAD_REQUEST, "존재하지 않는 역 id 입니다.");
+        }
         stationDao.deleteById(id);
     }
 }
